Copy self.headers dictionnary
Copy self.headers before editing, instead of creating only a new reference to it. This allows to re-use the same VectraClient instance to call multiple methods. 
I had a case where I created a new threat feed and then uploaded a STIX file the newly created feed. The issue is that since the "create_feed" function alters the "headers" attribute of the VectraClient instance to contain 'Content-Type': "application/json". When using the same VectraClient instance when calling "post_stix_file", the upload of a STIX file failed since we still sent the headers telling the API to expect JSON. 
Cheers
Aurélien